% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstTransitivity.R
\name{statsTransitivity}
\alias{statsTransitivity}
\title{Transitivity statistics for Dynamic Networks}
\usage{
statsTransitivity(X)
}
\arguments{
\item{X}{A \eqn{p \times p \times n} array of network's adjacency matrices over time.}
}
\value{
A list containing the statistics:
\itemize{
\item \code{U}: A \eqn{p \times p \times n} array of the normalized number of common neighbours
\item \code{V}: A \eqn{p \times p \times n} array of the normalized number of disjoint neighbours
}
}
\description{
This function calculates statistics which summarize the higher-order structure
of a dynamic network object: normalized number of common and disjoint neighbours
between each pair of nodes in each snapshot:
\deqn{U^{t-1}_{i,j} = \frac{1}{p-2} \sum_k X_{i,k}^{t-1} X_{k,j}^{t-1}}
\deqn{V^{t-1}_{i,j} = \frac{1}{p-2} \sum_k \{ X_{i,k}^{t-1}(1 - X_{k,j}^{t-1}) +  (1 - X_{i,k}^{t-1})X_{k,j}^{t-1} \}}
}
\examples{
p = 30; n = 20
xi = rep(0.7, p); eta = rep(0.8, p)
a = 30; b = 15

# Simulate data using simulate_transitivity function
simulated_data = simulate_transitivity(p, n, xi, eta, a, b)
X = simulated_data$X

# Calculate transitivity statistics
statsGlobal = statsTransitivity(X)


}
